requirements.txt
django==2.0.13
django-auditlog==0.4.5
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.7
djangorestframework==3.10.1
django-filter==2.1.0
django-filters==0.2.1
django-handyhelpers>=0.0.10
django-userextensions>=0.0.1
drf-dynamic-fields==0.3.1
djangorestframework-filters==0.11.1
